Pack question without reflection
